+++ src/Block/Element/FencedCode.php
@@ -219,12 +219,10 @@ class FencedCode extends AbstractBlock
      * @param Cursor $cursor
      * @param int    $currentLineNumber
      *
-     * @return $this
+     * @return bool
      */
-    public function setLastLineBlank(Cursor $cursor, $currentLineNumber)
+    public function shouldLastLineBeBlank(Cursor $cursor, $currentLineNumber)
     {
-        parent::setLastLineBlank($cursor, $currentLineNumber);
-
-        $this->lastLineBlank = false;
+        return false;
     }
 }